% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspsreports-package.R
\docType{package}
\name{cspsreports-package}
\alias{cspsreports}
\alias{cspsreports-package}
\title{cspsreports: Generate Civil Service People Survey Reports}
\description{
Render HTML and/or PDF reports given a specified organisational
    unit code or codes. Designed for use by designated analysts who have access
    to Civil Service People Survey data for their organisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/co-analysis/cspsreports/}
  \item \url{https://www.gov.uk/government/collections/civil-service-people-surveys}
  \item Report bugs at \url{https://github.com/co-analysis/cspsreports/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{csps@cabinetoffice.gov.uk}

}
\keyword{internal}
