% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generate A CSPS Report}
\usage{
generate_report(microdata, hierarchy, metadata, unit_codes, out_dir)
}
\arguments{
\item{microdata}{A data.frame. Your organisation's microdata file that
contains the responses of each respondent. Received as the file
\code{csps2019_flat.csv}.}

\item{hierarchy}{A data.frame. The lookup file that matches each
respondent to their unit code. Received in the file
\code{csps2019_hierarchy.csv}.}

\item{metadata}{A data.frame. The file that provides additional metadata
about each unit, such as headcount and name. Received in the file
\code{csps2019_hierarchy_metadata.csv}.}

\item{unit_codes}{A character vector of organisational unit codes in the
form \code{"XX0000"}.}

\item{out_dir}{Character. An existing directory to which the output will be
written.}
}
\value{
A rendered PDF file in the provided location.
}
\description{
Provide microdata and unit codes to generate a report from a
template. Designed for use by designated analysts who have access to
Civil Service People Survey data for their organisation.
}
\examples{
\dontrun{
    # Read files
    library(readr)
    micro_df <- read_csv("~/Documents/CO2019_flat.csv",)
    hier_df <- read_csv("~/Documents/CO2019_hierarchy.csv")
    meta_df <- read_csv("~/Documents/CO2019_hierarchy_metadata.csv")

    # Create a report for the specified unit codes
    generate_report(
        microdata = micro_df,
        hierarchy = hier_df,
        metadata = meta_df,
        unit_codes = c("CO0027", "CO0205", "CO0384"),
        out_dir = "~/Documents/reports"
   )
   }
}
