% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_microdata}
\alias{wrangle_microdata}
\title{Prepare Microdata For Reporting}
\usage{
wrangle_microdata(microdata, hierarchy, metadata, unit_codes)
}
\arguments{
\item{microdata}{A data.frame. Your organisation's microdata file that
contains the responses of each respondent. Received as the file
\code{csps2019_flat.csv}.}

\item{hierarchy}{A data.frame. The lookup file that matches each
respondent to their unit code. Received in the file
\code{csps2019_hierarchy.csv}.}

\item{metadata}{A data.frame. The file that provides additional metadata
about each unit, such as headcount and name. Received in the file
\code{csps2019_hierarchy_metadata.csv}.}

\item{unit_codes}{A character vector of organisational unit codes in the
form \code{"XX0000"}.}
}
\value{
A data.frame.
}
\description{
Merge microdata with hierarchy information and filter by a
provided vector of unit codes.
}
